application/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pika==1.3.2
paho-mqtt==1.6.1
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
librosa==0.10.1
soundfile==0.12.1
tensorflow==2.15.0
torch==2.1.1
torchaudio==2.1.1
opencv-python==4.8.1.78
Pillow==10.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
